% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomputeSupercells.R
\name{recomputeSupercells}
\alias{recomputeSupercells}
\title{Recompute supercells}
\usage{
recomputeSupercells(
  dt,
  sc_objects,
  markers,
  sample_colname,
  cell_id_colname,
  gam = 20
)
}
\arguments{
\item{dt}{\link{data.table} object containing the cytometry data.
Rows represent cells, columns represent markers.
If this is not a \link{data.table} object, the function will warn you about it,
and then try to convert it to a \link{data.table} object.}

\item{sc_objects}{The \code{supercell_object} returned by \link{runSuperCellCyto} function.}

\item{markers}{character vector specifying the markers in \code{dt}.}

\item{sample_colname}{character specifying the column in \code{dt} that denotes
the sample of a cell.}

\item{cell_id_colname}{character specifying the column in \code{dt} that denotes
the unique ID of a cell.}

\item{gam}{numeric specifying the gamma value to be used by SuperCell.}
}
\value{
A list with the following components:
\describe{
\item{\code{supercell_expression_matrix}:}{A \link{data.table} containing
the marker expression of all the supercells.
These are computed by taking the average marker expression of the cells
captured by each supercell.}
\item{\code{supercell_cell_map}:}{A \link{data.table} showing which cell is
captured by which supercell. This is very useful if you intend to work out
which supercell captures which cell.}
}
}
\description{
Given a supercell object, recompute the supercell granularity using a different
gamma value.
Gamma value controls the number of supercells generated.
The smaller the value, the more supercells you get, and vice versa.
}
\examples{
set.seed(42)
cyto_dat <- simCytoData(10, rep(1000, 3))
markers <- paste0("Marker_", seq_len(10))
out_gam20 <- runSuperCellCyto(
    dt = cyto_dat,
    markers = markers,
    sample_colname = "Sample",
    cell_id_colname = "Cell_Id",
    gam = 20
)
recomputed_sc <- recomputeSupercells(
    dt = cyto_dat,
    sc_objects = out_gam20$supercell_object,
    markers = markers,
    sample_colname = "Sample",
    cell_id_colname = "Cell_Id",
    gam = 50
)

}
\author{
Givanna Putri
}
