% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSuperCellCyto.R
\name{runSuperCellCyto}
\alias{runSuperCellCyto}
\title{Run SuperCell for cytometry data}
\usage{
runSuperCellCyto(
  dt,
  markers,
  sample_colname,
  cell_id_colname,
  gam = 20,
  k_knn = 5,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{dt}{A \link{data.table} object containing the cytometry data.
Rows represent cells, columns represent markers.}

\item{markers}{A character vector specifying the markers to run SuperCell on.}

\item{sample_colname}{String specifying the column in \code{dt} that denotes
the sample of a cell.}

\item{cell_id_colname}{String specifying the column in \code{dt} that denotes
the unique ID of a cell.}

\item{gam}{Numeric scalar specifying the gamma value to be used by SuperCell.}

\item{k_knn}{Numeric scalar specifying the k value to be used by SuperCell's knn.}

\item{BPPARAM}{\linkS4class{BiocParallelParam} object specifying the
configuration parameters for parallel execution.
Default to \linkS4class{SerialParam}, i.e., not parallelisation to be used.

#' @section Details about SuperCell:
We used SuperCell's SCimplify method to generate supercells for cytometry data.
To install SuperCell package, please use \link[remotes]{install_github} as
it is only available on github:
\code{remotes::install_github("GfellerLab/SuperCell")}.

By default, all the markers specified in \code{markers} parameter are used
to compute PCA,and that the marker expressions are \emph{not} scaled when
computing PCA.
\code{irlba} is not used to calculate PCA as cytometry data only have a
handful of features (markers) in general.
Number of PCs are set to 10, default in SCimplify.

\code{gam} and \code{k_knn} are passed on as it is to indicate the graining
level of supercells and the k value used to compute the single-cell
kNN network.
Actual (not approximate) kNN network is created, and walktrap algorithm was
used to detect the supercells from the kNN network.

Cells from each sample are processed independent of those from different samples.
To differentiate the same supercell ID across multiple samples,
the supercell IDs are prepended with the sample ID.
For instance, supercell 1 from sample 1 will be named SuperCell_1_Sample_1,
whereas supercell 1 from sample 2 will be named SuperCell_2_Sample_1.

If none of the above make sense to you, please read Bilous et.al, 2022
manuscript to see how SuperCell works.}
}
\value{
\code{runSuperCellCyto} will return a list with the following components:
\describe{
\item{\code{supercell_object}:}{A list containing a list returned by
SCimplify function for each sample.}
\item{\code{supercell_expression_matrix}:}{A \link{data.table} containing
the marker expression of all the supercells.}
\item{\code{supercell_cell_map}:}{A \link{data.table} containing the
supercell ID of all the cells in the data.}
}

Each supercell's marker expression in \code{supercell_expression_matrix} is
computed based on the average marker expression of all the cells captured
by the supercell.
}
\description{
Run SuperCell on cytometry data stored as \link{data.table} object.
The vanilla SuperCell algorithm is provided by the SuperCell R package
(Bilous et.al, 2022).
We have enhanced it by adding the capacity to "supercell" multiple samples
in parallel
through the use of \link{BiocParallel}, and by adding support to "supercell"
cytometry data.
See Details below on how SuperCell was used.
}
\examples{
# Simulate some data
set.seed(42)
cyto_dat <- simCytoData()

# Setup the columns designating the markers, samples, and cell IDs
marker_col <- paste0("Marker_", seq_len(nmarkers))
sample_col <- "Sample"
cell_id_col <- "Cell_Id"

supercell_dat <- runSuperCellCyto(cyto_dat, marker_col, 
sample_col, cell_id_col)


}
\author{
Givanna Putri
}
