% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSuperCellCyto.R
\name{runSuperCellCyto}
\alias{runSuperCellCyto}
\title{Run SuperCell for cytometry data}
\usage{
runSuperCellCyto(
  dt,
  markers,
  sample_colname,
  cell_id_colname,
  gam = 20,
  k_knn = 5,
  seed = 42,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{dt}{A data.table object containing a cytometry data
where a row is a cell and a column is a marker.}

\item{markers}{A character vector specifying the marker to run SuperCell on.}

\item{sample_colname}{String specifying the column in \code{dt} that denotes
the sample of a cell.}

\item{cell_id_colname}{String specifying the column in \code{dt} that denotes
the unique ID of a cell.}

\item{gam}{Numeric scalar specifying the gamma value to be used by SuperCell.}

\item{k_knn}{Numeric scalar specifying the k value to be used by SuperCell's knn.}

\item{seed}{Numeric scalar specifying the seed for random number generator.}

\item{BPPARAM}{\linkS4class{BiocParallelParam} object specifying the configuration parameters for parallel execution.
Default to \linkS4class{SerialParam}, i.e., not parallelisation to be used.}
}
\description{
Run SuperCell for cytometry data
}
\author{
Givanna Putri
}
