% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSuperCellCyto.R
\name{runSuperCellCyto}
\alias{runSuperCellCyto}
\title{Run SuperCell for cytometry data}
\usage{
runSuperCellCyto(
  dt,
  markers,
  sample_colname,
  cell_id_colname,
  gam = 20,
  k_knn = 5,
  seed = 42,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{dt}{A \link{data.table} object containing the cytometry data.
Rows represent cells, columns represent markers.}

\item{markers}{A character vector specifying the markers to run SuperCell on.}

\item{sample_colname}{String specifying the column in \code{dt} that denotes
the sample of a cell.}

\item{cell_id_colname}{String specifying the column in \code{dt} that denotes
the unique ID of a cell.}

\item{gam}{Numeric scalar specifying the gamma value to be used by SuperCell.}

\item{k_knn}{Numeric scalar specifying the k value to be used by SuperCell's knn.}

\item{seed}{Numeric scalar specifying the seed for random number generator.}

\item{BPPARAM}{\linkS4class{BiocParallelParam} object specifying the configuration parameters for parallel execution.
Default to \linkS4class{SerialParam}, i.e., not parallelisation to be used.}
}
\value{
\code{runSuperCellCyto} will return a list with the following components:
\describe{
\item{\code{supercell_object}:}{A list containing a list returned by \link{SCimplify} function for each sample.}
\item{\code{supercell_expression_matrix}:}{A \link{data.table} containing the marker expression of all the supercells.}
\item{\code{supercell_cell_map}:}{A \link{data.table} containing the supercell ID of all the cells in the data.}
}

Each supercell's marker expression in \code{supercell_expression_matrix} is computed based on the average marker expression of all the cells captured by the supercell.
}
\description{
Run \link{SuperCell} on cytometry data stored as \link{data.table} object.
}
\author{
Givanna Putri
}
