% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCytoData.R
\name{simCytoData}
\alias{simCytoData}
\title{Simulates cytometry data}
\usage{
simCytoData(nmarkers = 10, ncells = 10000, nsample = 2)
}
\arguments{
\item{nmarkers}{Numeric scalar specifying number of markers to simulate.}

\item{ncells}{Numeric scalar specifying number of cells to simulate per sample.}

\item{nsample}{Numeric scalar specifying number of samples to simulate.}
}
\value{
A \link{data.table} containing the simulated data.
Rows are cells, columns are markers.
}
\description{
Simulates some cytometry data for use in testing or documenting functions
which require some cytometry data.
Please run \code{set.seed} before running the function if you want to ensure
reproducibility.
}
\examples{
set.seed(42)
cyto_dat <- simCytoData()
head(cyto_dat) 

}
\author{
Givanna Putri
}
